# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Bmstu Open IT Platform",
        "title": "Forecast system",
        "contact": {
            "name": "API Support",
            "url": "https://vk.com/b.larkin",
            "email": "borislarkin18@mail.ru"
        },
        "license": {
            "name": "AS IS (NO WARRANTY)"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/prediction/delete/{id}": {
            "delete": {
                "description": "Method sets prediction`s status to \"deleted\" without actually removing it from the db model",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Predictions"
                ],
                "summary": "Delete specified prediction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Prediction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/prediction/edit/{id}": {
            "put": {
                "description": "very very friendly response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Predictions"
                ],
                "summary": "Edit specified prediction`s prediction amount \u0026 window",
                "parameters": [
                    {
                        "description": "New prediction data",
                        "name": "prediction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.EditPredReq"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Prediction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.PredictionWithForecasts"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/prediction/finish/{id}": {
            "put": {
                "description": "Can be ended with statuses: [\"denied\", \"completed\"]",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Predictions"
                ],
                "summary": "Finish specified prediction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Prediction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status to be set",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.PredictionWithForecasts"
                        }
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
        "/prediction/form/{id}": {
            "put": {
                "description": "very very friendly response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Predictions"
                ],
                "summary": "Form specified prediction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Prediction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.PredictionWithForecasts"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/prediction/{id}": {
            "get": {
                "description": "very very friendly response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Predictions"
                ],
                "summary": "Display a prediction and its forecasts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Prediction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.PredictionWithForecasts"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/predictions": {
            "get": {
                "description": "very very friendly response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Predictions"
                ],
                "summary": "Show all predictions made for current user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prediction status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Earliest date created filter: YYYY-Mon-DD",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Latest date created filter: YYYY-Mon-DD",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.Predictions"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.ForecastResponseWithFlags": {
            "type": "object",
            "required": [
                "color",
                "descr",
                "ext_desc",
                "forecast_id",
                "image",
                "input",
                "measure_type",
                "result",
                "short_title",
                "title"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "descr": {
                    "type": "string"
                },
                "ext_desc": {
                    "type": "string"
                },
                "forecast_id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "input": {
                    "type": "string"
                },
                "measure_type": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "short_title": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "ds.PredictionWithForecasts": {
            "type": "object",
            "required": [
                "forecasts",
                "prediction"
            ],
            "properties": {
                "forecasts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.ForecastResponseWithFlags"
                    }
                },
                "prediction": {
                    "$ref": "#/definitions/ds.Predictions"
                }
            }
        },
        "ds.Predictions": {
            "type": "object",
            "properties": {
                "date_completed": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "date_formed": {
                    "type": "string"
                },
                "prediction_amount": {
                    "type": "integer"
                },
                "prediction_id": {
                    "type": "integer"
                },
                "prediction_window": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handler.EditPredReq": {
            "type": "object",
            "required": [
                "prediction_amount",
                "prediction_window"
            ],
            "properties": {
                "prediction_amount": {
                    "type": "integer"
                },
                "prediction_window": {
                    "type": "integer"
                }
            }
        },
...
    }
}